+++ lib/src/widget/smart_dialog_widget.dart
@@ -2,11 +2,13 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_smart_dialog/src/data/base_controller.dart';
+import 'package:flutter_smart_dialog/src/util/view_utils.dart';
 import 'package:flutter_smart_dialog/src/widget/animation/fade_animation.dart';
 import 'package:flutter_smart_dialog/src/widget/animation/scale_animation.dart';
 import 'package:flutter_smart_dialog/src/widget/animation/slide_animation.dart';
 
 import '../config/enum_config.dart';
+import '../data/animation_param.dart';
 import 'animation/mask_animation.dart';
 import 'helper/mask_event.dart';
 
@@ -21,6 +23,7 @@ class SmartDialogWidget extends StatefulWidget {
     required this.animationTime,
     required this.useAnimation,
     required this.animationType,
+    required this.animationBuilder,
     required this.maskColor,
     required this.maskWidget,
     required this.maskTriggerType,
@@ -30,7 +33,7 @@ class SmartDialogWidget extends StatefulWidget {
   final Widget child;
 
   ///widget controller
-  final SmartDialogController controller;
+  final SmartDialogWidgetController controller;
 
   /// 点击遮罩
   final VoidCallback onMask;
@@ -51,6 +54,9 @@ class SmartDialogWidget extends StatefulWidget {
   /// 仅仅针对中间位置的控件
   final SmartAnimationType animationType;
 
+  /// 自定义动画
+  final AnimationBuilder? animationBuilder;
+
   /// 遮罩颜色
   final Color maskColor;
 
@@ -68,6 +74,7 @@ class _SmartDialogWidgetState extends State<SmartDialogWidget>
     with TickerProviderStateMixin {
   AnimationController? _maskController;
   late AnimationController _bodyController;
+  AnimationParam? _animationParam;
 
   @override
   void initState() {
@@ -90,6 +97,10 @@ class _SmartDialogWidgetState extends State<SmartDialogWidget>
       _bodyController.value = 0;
       _bodyController.forward();
     }
+
+    ViewUtils.addSafeUse(() {
+      _animationParam?.onForward?.call();
+    });
 
     //bind controller
     widget.controller._bind(this);
@@ -126,6 +137,17 @@ class _SmartDialogWidgetState extends State<SmartDialogWidget>
 
   Widget _buildBodyAnimation() {
     var child = widget.child;
+    if (widget.animationBuilder != null) {
+      return widget.animationBuilder!.call(
+        _bodyController,
+        child,
+        _animationParam = AnimationParam(
+          alignment: widget.alignment,
+          animationTime: widget.animationTime,
+        ),
+      );
+    }
+
     var type = widget.animationType;
     Widget fade = FadeAnimation(controller: _bodyController, child: child);
     Widget scale = ScaleAnimation(controller: _bodyController, child: child);
@@ -163,8 +185,11 @@ class _SmartDialogWidgetState extends State<SmartDialogWidget>
 
     _maskController!.reverse();
     _bodyController.reverse();
+    _animationParam?.onDismiss?.call();
+
+    var awaitTime = _bodyController.duration ?? widget.animationTime;
     if (widget.useAnimation) {
-      await Future.delayed(widget.animationTime);
+      await Future.delayed(awaitTime);
     }
   }
 
@@ -178,7 +203,7 @@ class _SmartDialogWidgetState extends State<SmartDialogWidget>
   }
 }
 
-class SmartDialogController extends BaseController {
+class SmartDialogWidgetController extends BaseController {
   _SmartDialogWidgetState? _state;
 
   void _bind(_SmartDialogWidgetState _state) {
@@ -187,13 +212,6 @@ class SmartDialogController extends BaseController {
 
   @override
   Future<void> dismiss() async {
-    try {
-      await _state?.dismiss();
-    } catch (e) {
-      print("-------------------------------------------------------------");
-      print("SmartDialog error: ${e.toString()}");
-      print("-------------------------------------------------------------");
-    }
-    _state = null;
+    await _state?.dismiss();
   }
 }